+++ Game2/Src/Game2/EntityStats.cpp
@@ -8,7 +8,7 @@ namespace Engine
     float Stats::attack_timer = 0.0f;
 
     Stats::Stats(float initialHealth, float initialAttackPower)
-        : health(initialHealth), attackPower(initialAttackPower) {}
+        : health(static_cast<int>(initialHealth)), attackPower(static_cast<int>(initialAttackPower)) {}
 
     int Stats::getHealth() const
     {
@@ -22,12 +22,12 @@ namespace Engine
 
     void Stats::setHealth(float newHealth)
     {
-        health = newHealth;
+        health = static_cast<int>(newHealth);
     }
 
     void Stats::setAttackPower(float newAttackPower)
     {
-        attackPower = newAttackPower;
+        attackPower = static_cast<int>(newAttackPower);
     }
 
     void Stats::displayStats() const
